+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.6.3
 - Fix incompatibility with Slayer's Beasts, remove duplicate feature from Coniferous Forest(s)
+- Fix Tall Beach and Prairie Grass dropping two tall grass instead of normal grass
 
 # 1.6.2
 - Fix White Puffball not growing over time